fix(zle): zleread never popped the buffer stack, so push-line and run-help destroyed the line
Every widget that stashes the current line pushed it and nothing ever brought
it back. `zle_misc.rs:602` ports c:411's `zpushnode(bufstack, …)` and saves the
cursor at c:412, but `zleread` omitted c:Src/Zle/zle_main.c:1297-1312 entirely:

    if ((s = getlinknode(bufstack))) {
        setline(s, ZSL_TOEND);
        if (stackcs != -1) { zlecs = stackcs; stackcs = -1;
                             if (zlecs > zlell) zlecs = zlell; CCLEFT(); }
        if (stackhist != -1) { histline = stackhist; stackhist = -1; }
        handleundo();
    }

Measured against /opt/homebrew/bin/zsh through a PTY, emacs bindings:

    push-line (ESC-q) then CR
      zsh    READY% READY% print first      before  READY% READY% READY%
    run-help (M-h) then X
      zsh    READY% echoo testX             before  READY% X

`STACKCS` also had to change type. It was `AtomicUsize` initialised 0, which
cannot represent the -1 sentinel `if (stackcs != -1)` tests; C declares
`int stackhist, stackcs` (c:zle_main.c:162) and initialises both to -1
(c:2257). Without the signed type the cursor restore fires when it must not.

`print -z` and `printf -z` pushed with `.push()` where `zpushnode` inserts at
the list HEAD (zsh.h:591). Every consumer — zleread's restore (c:1297),
`read -z` (c:builtin.c:6762), `get-line` (c:zle_hist.c:901) — takes the head
back off, so appending made the stack FIFO and two queued lines returned in the
wrong order.

Verified unchanged on the shared consumers, byte-identical to zsh:
`print -z`/`read -z` round trip; two queued lines coming back newest-first;
`read -z` on an empty stack still rc=1; `print -z` preserving an embedded
space; and M-b (backward-word) as a widget control.

tests/parity/zle_bufstack_parity.rs pins push-line's line AND cursor, two
stacked push-lines, run-help, print -z, and accept-and-hold. zle_main 66,
zle_misc 108, zle_hist 70, zle_utils 83, bufstack parity 5 — all pass.

+++ src/ported/zle/zle_misc.rs
@@ -600,7 +600,7 @@ pub fn acceptandhold() -> i32 {
     let zlell = ZLELL.load(SeqCst);
     let line: String = ZLELINE.lock().unwrap().iter().take(zlell).collect();
     BUFSTACK.lock().unwrap().insert(0, line); // c:411 zpushnode
-    STACKCS.store(ZLECS.load(SeqCst), SeqCst); // c:412
+    STACKCS.store(ZLECS.load(SeqCst) as i32, SeqCst); // c:412
     DONE.store(1, SeqCst); // c:413
     0 // c:414
 }
@@ -3979,7 +3979,7 @@ mod tests {
         let r = acceptandhold();
         assert_eq!(r, 0);
         assert_eq!(DONE.load(SeqCst), 1);
-        assert_eq!(STACKCS.load(SeqCst), 4, "STACKCS must capture ZLECS");
+        assert_eq!(STACKCS.load(SeqCst), 4i32, "STACKCS must capture ZLECS");
         let bs = BUFSTACK.lock().unwrap();
         assert_eq!(bs[0], "test line", "BUFSTACK[0] must be the pushed line");
     }
